requirements.txt
asgiref==3.8.1
Django==4.2.19
django-cors-headers==4.7.0
django-tinymce==4.1.0
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
    